+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
+## Unreleased – Unreleased
+
+## Additions
+
+- Adds `lintjs` task.
+- Adds `--quiet` CLI flag to Gruntfile to suppress warnings in linter.
+
+## Changes
+
+- Supersedes `eslint` task by `lintjs`.
+
+
 ## 0.7.0 - 2015-05-19
 
 ### Added